+++ app/src/main/java/com/phisit/weatherforecast/domain/model/WeatherModel.kt
@@ -10,14 +10,14 @@ data class WeatherModel(
 
 data class CurrentModel(
     val clouds: Int = 0,
-    val dewPoint: Double = 0.0,
+    val dewPoint: Int = 0,
     val dt: Int = 0,
     val feelsLike: Double = 0.0,
     val humidity: Int = 0,
     val pressure: Int = 0,
     val sunrise: Int = 0,
     val sunset: Int = 0,
-    val temp: Double = 0.0,
+    val temp: Int = 0,
     val uvi: Double = 0.0,
     val visibility: Int = 0,
     val weather: List<WeatherDetailModel> = listOf(),
